Add helper method for handling purposes
